requirements.txt
requests
gitpython
tqdm
transformers
urllib3
